fix: 第二波编译错误修复 — path包导入, PermissionKind枚举, skills导入, models空安全, riverpod
android_tools: 在入口文件添加 path 包 import, 解决子模块 p.undefined 错误

android_tools_advanced: toPx 返回类型 num→double.toDouble()

skill_tools: 添加 agent_constants.dart 导入 (kSkillDescPreviewMax)

skills: 移除 android_tools_advanced.dart 直接导入, 改为通过 android_tools.dart 入口

android_automation_service: PermissionKind 枚举移到类外, 添加默认 return 分支, notificationListenerEnabled→Granted

models: sizeMb/ramMb 去掉多余的 != null 判断, 直接 as num? 可空

pubspec: 添加 flutter_riverpod ^2.5.0 依赖

test_utils: 添加 dart:async 导入 (TimeoutException)

agent_reliability_benchmark: 修复 v@#$% 字符串转义为 v@#\$%

rag_test: CachedDoc.lastAccessed 从 null 改为 DateTime(2024,1,1)

+++ lib/agent/android_tools/android_tools_advanced.dart
@@ -1163,7 +1163,7 @@ Tool _visionAnalyzeRegionTool(
         final res = await s.screenResolution();
         final w = (res?[0] ?? 1080).toDouble();
         final h = (res?[1] ?? 2400).toDouble();
-        double toPx(num v, double max) => v < 1 ? v * max : v;
+        double toPx(num v, double max) => v < 1 ? v * max : v.toDouble();
         final x = toPx((args['x'] as num?)?.toDouble() ?? 0, w);
         final y = toPx((args['y'] as num?)?.toDouble() ?? 0, h);
         final rw = toPx((args['width'] as num?)?.toDouble() ?? w, w);
